glopfile_ffi_bindings

Module bindings

Generate foreign language bindings for a uniffi component.

This module contains all the code for generating foreign language bindings, along with some helpers for executing foreign language scripts or tests.

Modules

python

Structs

GenerateOptions

SwiftBindingsOptions

Enums

TargetLanguage

Functions

generate

Generate bindings

generate_swift_bindings

Generate Swift bindings (specialized version)

generate_with_bindgen_paths

Generate bindings with custom bindgen paths.